split contexts module into individual files
